git commit after applied size threshold print num to sync and synced

+++ file-tree-merge/src/apply_change.rs
@@ -27,11 +27,18 @@ pub fn apply(
     if changes.is_empty() {
         return Ok(());
     }
-    let ctr = AtomicU64::new(0);
+    println!(
+        "{}",
+        format!("Applying {} changes...", changes.len()).cyan()
+    );
+    let total = AtomicU64::new(0);
+    let synced = AtomicU64::new(0);
+    let applied_size_since_commit = AtomicU64::new(0);
     let items_path1 = Arc::new(Mutex::new(items_path1));
     let items_path2 = Arc::new(Mutex::new(items_path2));
     let mut to_process = vec![];
     let batch_size = 1000;
+    let commit_after_size_bytes = 64 * 1024 * 1024;
     let git_lock = Arc::new(Mutex::new(()));
     for (change, path) in changes {
         to_process.push((change.clone(), path.to_string()));
@@ -52,14 +59,27 @@ pub fn apply(
                         dry_run,
                         checksum,
                         crc,
-                        &ctr,
+                        &total,
                         git_lock.clone(),
                         batch_size,
+                        &synced,
+                        &applied_size_since_commit,
+                        commit_after_size_bytes,
                     )
                 })
                 .collect();
             for e in res {
                 if let Err(err) = e {
+                    println!(
+                        "{}",
+                        format!(
+                            "Synced {}/{}",
+                            synced.load(Ordering::SeqCst),
+                            total.load(Ordering::SeqCst)
+                        )
+                        .green()
+                        .bold()
+                    );
                     Err(err)?
                 }
             }
@@ -82,17 +102,43 @@ pub fn apply(
                 dry_run,
                 checksum,
                 crc,
-                &ctr,
+                &total,
                 git_lock.clone(),
                 batch_size,
+                &synced,
+                &applied_size_since_commit,
+                commit_after_size_bytes,
             )
         })
         .collect();
-    println!("processed {}", ctr.load(Ordering::SeqCst));
     for e in res {
         if let Err(err) = e {
+            if !dry_run {
+                println!(
+                    "{}",
+                    format!(
+                        "Synced {}/{}",
+                        synced.load(Ordering::SeqCst),
+                        total.load(Ordering::SeqCst)
+                    )
+                    .green()
+                    .bold()
+                );
+            }
             Err(err)?
         }
+    }
+    if !dry_run {
+        println!(
+            "{}",
+            format!(
+                "Synced {}/{}",
+                synced.load(Ordering::SeqCst),
+                total.load(Ordering::SeqCst)
+            )
+            .green()
+            .bold()
+        );
     }
     git_add(&path1_repo.join(TREE_DIR), ".")?;
     git_commit(path1_repo)?;
@@ -135,8 +181,12 @@ fn process(
     ctr: &AtomicU64,
     git_lock: Arc<Mutex<()>>,
     batch_size: usize,
+    synced: &AtomicU64,
+    applied_size_since_commit: &AtomicU64,
+    commit_after_size_bytes: i32,
 ) -> Result<()> {
     let path2 = path2_mnt.join(path);
+    ctr.fetch_add(1, Ordering::SeqCst);
     match change.clone() {
         Change::Add | Change::Modify => {
             if ctr.load(Ordering::SeqCst) % batch_size as u64 == 0 {
@@ -145,9 +195,6 @@ fn process(
                 } else {
                     println!("{} '{}'", change.to_string().blue(), path.blue());
                 }
-            }
-            if dry_run {
-                return Ok(());
             }
             // check if it's the same as in dst
             let mut add = true;
@@ -159,6 +206,9 @@ fn process(
                 }
             }
             if add {
+                if dry_run {
+                    return Ok(());
+                }
                 fs::create_dir_all(path2.parent().unwrap())?;
                 fs::copy(path1_mnt.join(&path), path2.clone())?;
                 File::set_times(&File::open(path2.clone())?, path1_item.times)?;
@@ -172,6 +222,8 @@ fn process(
                     );
                     anyhow::bail!("CRC check failed for `{path}` after transfer");
                 }
+                synced.fetch_add(1, Ordering::SeqCst);
+                applied_size_since_commit.fetch_add(path1_item.size, Ordering::SeqCst);
             } else if ctr.load(Ordering::SeqCst) % batch_size as u64 == 0 {
                 println!(
                     "{}",
@@ -183,27 +235,28 @@ fn process(
             if ctr.load(Ordering::SeqCst) % batch_size as u64 == 0 {
                 println!("{} '{}'", change.to_string().red(), path.red().bold());
             }
-            if dry_run {
-                return Ok(());
-            }
             if path2.exists() {
+                if dry_run {
+                    return Ok(());
+                }
                 fs::remove_file(path2.clone())?;
                 File::open(path2.parent().unwrap())?.sync_all()?;
             } else if ctr.load(Ordering::SeqCst) % batch_size as u64 == 0 {
                 println!("{}", "  skip, not present in path2".yellow());
             }
+            synced.fetch_add(1, Ordering::SeqCst);
         }
         Change::Rename(old_path) => {
             if ctr.load(Ordering::SeqCst) % batch_size as u64 == 0 {
                 println!("{} '{}'", change.to_string().magenta(), path.magenta());
             }
-            if dry_run {
-                return Ok(());
-            }
             let guard = items_path1.lock().unwrap();
             let path1_item = guard.get(path).unwrap();
             // todo: compare if old file hash in src is same as old file hash in dst
             if path2_mnt.join(&old_path).exists() {
+                if dry_run {
+                    return Ok(());
+                }
                 fs::create_dir_all(path2.parent().unwrap())?;
                 fs::rename(path2_mnt.join(&old_path), path2.clone())?;
                 File::set_times(&File::open(path2.clone())?, path1_item.times)?;
@@ -211,6 +264,9 @@ fn process(
                 File::open(path2.parent().unwrap())?.sync_all()?;
             } else {
                 println!("{}", format!("  cannot R '{old_path}' -> '{path}', old file not present in path2. Will copy instead from path1 to the new destination").yellow());
+                if dry_run {
+                    return Ok(());
+                }
                 fs::create_dir_all(path2_mnt.join(path).parent().unwrap())?;
                 fs::copy(path1_mnt.join(path), path2.clone())?;
                 File::set_times(&File::open(path2.clone())?, path1_item.times)?;
@@ -224,19 +280,21 @@ fn process(
                     );
                     anyhow::bail!("CRC check failed for `{path}` after transfer");
                 }
+                applied_size_since_commit.fetch_add(path1_item.size, Ordering::SeqCst);
             }
+            synced.fetch_add(1, Ordering::SeqCst);
         }
         Change::Copy(old_path) => {
             if ctr.load(Ordering::SeqCst) % batch_size as u64 == 0 {
                 println!("{} '{}'", change.to_string().blue(), path.blue());
             }
-            if dry_run {
-                return Ok(());
-            }
             let guard = items_path1.lock().unwrap();
             let path1_item = guard.get(path).unwrap();
             // todo: compare if old file hash in src is same as old file hash in dst
             if path2_mnt.join(&old_path).exists() {
+                if dry_run {
+                    return Ok(());
+                }
                 fs::create_dir_all(path2.clone().parent().unwrap())?;
                 fs::copy(path2_mnt.join(&old_path), path2.clone())?;
                 File::set_times(&File::open(path2.clone())?, path1_item.times)?;
@@ -244,6 +302,9 @@ fn process(
                 File::open(path2.parent().unwrap())?.sync_all()?;
             } else {
                 println!("{}", format!("  cannot C '{old_path}' -> '{path}', old file not present in path2. Will copy instead from path1 to the new destination").yellow());
+                if dry_run {
+                    return Ok(());
+                }
                 fs::create_dir_all(path2.parent().unwrap())?;
                 fs::copy(path1_mnt.join(path), path2.clone())?;
             }
@@ -255,13 +316,16 @@ fn process(
                 );
                 anyhow::bail!("CRC check failed for `{path}` after transfer");
             }
+            synced.fetch_add(1, Ordering::SeqCst);
+            applied_size_since_commit.fetch_add(path1_item.size, Ordering::SeqCst);
         }
     }
     let _guard = git_lock.lock().unwrap();
     git_add(&path1_repo.join(TREE_DIR), path)?;
-    // todo: make in based on file size
-    if ctr.fetch_add(1, Ordering::SeqCst) % 100 == 0 {
+    if applied_size_since_commit.load(Ordering::SeqCst) > commit_after_size_bytes as u64 {
+        println!("{}", "Checkpointing applied changes...".cyan());
         git_commit(path1_repo)?;
+        applied_size_since_commit.store(0, Ordering::SeqCst);
     }
     Ok(())
 }
